修复:附魔系统审查发现的问题
- 消化速率倍数钳制为非负,防止 multiply <= -1 时消化反向(水/食物灌入)
- get_max_hp_add 改返回 double,保留 add 小数精度,与 multiply 对称
- limb_score per-part 修饰跳过不提供该 score 的部位(mod==0),避免凭空加分
- manip 得分的 per-part 修饰在自然上限钳制后应用,增强可突破 limb_score_max
- enchant_cache::load 对缺省/显式空 part 统一走全局分支(is_empty 判定)
- MAGIC.md 修正 math 表达式支持范围措辞,补充零值部位与 manip 钳制行为说明
- 新增 test_LIMB_SCORE_ZERO_BP_ench_item 及零值部位跳过断言

+++ src/magic_enchantment.h
@@ -377,7 +377,7 @@ class enchant_cache : public enchantment
         double get_armor_multiply( const damage_type_id &value ) const;
         double get_encumbrance_multiply( const bodypart_str_id &value ) const;
         double get_extra_damage_multiply( const damage_type_id &value ) const;
-        int get_max_hp_add( const bodypart_str_id &value ) const;
+        double get_max_hp_add( const bodypart_str_id &value ) const;
         double get_max_hp_multiply( const bodypart_str_id &value ) const;
         // global limb score modifiers (apply to the final aggregated score)
         double get_limb_score_add( const limb_score_id &score ) const;
